examples: jwt desktop swift toolkit support for iOS has been added.
